+++ jest.config.js
@@ -1,5 +1,14 @@
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+const globals = {
+  'ts-jest': {
+    diagnostics: {
+      // https://github.com/kulshekhar/ts-jest/issues/1647#issuecomment-832577036
+      pathRegex: /\.(test)\.tsx$/,
+    },
+  },
+};
+
+const project = {
   // TypeScript files (.ts, .tsx) will be transformed by ts-jest to CommonJS syntax, and JavaScript files (.js, jsx) will be transformed by babel-jest.
   preset: 'ts-jest/presets/js-with-babel',
   testEnvironment: 'jsdom',
@@ -12,18 +21,47 @@ module.exports = {
     '@utils(.*)$': '<rootDir>/webapp/javascript/util/$1',
     '@models(.*)$': '<rootDir>/webapp/javascript/models/$1',
     '@ui(.*)$': '<rootDir>/webapp/javascript/ui/$1',
+    '@pyroscope/redux(.*)$': '<rootDir>/webapp/javascript/redux/$1',
+    '@pyroscope/services(.*)$': '<rootDir>/webapp/javascript/services/$1',
   },
   transform: {
     '\\.module\\.(css|scss)$': 'jest-css-modules-transform',
     '\\.(css|scss)$': 'jest-css-modules-transform',
     '\\.svg$': 'svg-jest',
   },
+  transformIgnorePatterns: [
+    // force us to transpile these dependenciesmyth
+    'node_modules/(?!(true-myth|react-notifications-component)/)',
+  ],
   globals: {
     'ts-jest': {
+      tsconfig: `tsconfig.test.json`,
       diagnostics: {
         // https://github.com/kulshekhar/ts-jest/issues/1647#issuecomment-832577036
         pathRegex: /\.(test)\.tsx$/,
       },
     },
   },
+};
+
+module.exports = {
+  // https://github.com/kulshekhar/ts-jest/issues/1648#issuecomment-820860089
+  // We create multiple projects so that we can slowly change tsconfig to be more strict
+  // Without affecting existing code
+  projects: [
+    {
+      ...project,
+      globals: {
+        ...project.globals,
+        'ts-jest': {
+          ...project.globals['ts-jest'],
+          tsconfig: 'webapp/javascript/services/tsconfig.json',
+        },
+      },
+      testMatch: ['<rootDir>/webapp/javascript/services/**/*.(spec|test).*'],
+    },
+    {
+      ...project,
+    },
+  ],
 };