+++ scripts/webpack/webpack.size-limit.ts
@@ -1,0 +1,16 @@
+import { merge } from 'webpack-merge';
+import path from 'path';
+import prod from './webpack.prod';
+
+module.exports = merge(prod, {
+  output: {
+    publicPath: '',
+    path: path.resolve(__dirname, '../../webapp/public/assets'),
+
+    // build WITHOUT hash
+    // otherwise size-limit won't know what to compare to
+    // https://github.com/andresz1/size-limit-action/issues/47
+    filename: '[name].js',
+    clean: true,
+  },
+});