+++ webapp/javascript/redux/actions.js
@@ -19,17 +19,39 @@ import {
   SET_RIGHT_UNTIL,
   RECEIVE_COMPARISON_APP_DATA,
   REQUEST_COMPARISON_APP_DATA,
-  REQUEST_PYRESCOPE_APP_DATA,
-  RECEIVE_PYRESCOPE_APP_DATA,
+  CANCEL_COMPARISON_APP_DATA,
+  REQUEST_PYROSCOPE_APP_DATA,
+  RECEIVE_PYROSCOPE_APP_DATA,
+  CANCEL_PYROSCOPE_APP_DATA,
   REQUEST_COMPARISON_DIFF_APP_DATA,
   RECEIVE_COMPARISON_DIFF_APP_DATA,
+  CANCEL_COMPARISON_DIFF_APP_DATA,
   REQUEST_COMPARISON_TIMELINE,
   RECEIVE_COMPARISON_TIMELINE,
-  SET_FILE,
-  SET_LEFT_FILE,
-  SET_RIGHT_FILE,
+  SET_ADHOC_FILE,
+  SET_ADHOC_LEFT_FILE,
+  SET_ADHOC_RIGHT_FILE,
+  REQUEST_ADHOC_PROFILES,
+  RECEIVE_ADHOC_PROFILES,
+  CANCEL_ADHOC_PROFILES,
+  SET_ADHOC_PROFILE,
+  REQUEST_ADHOC_PROFILE,
+  RECEIVE_ADHOC_PROFILE,
+  CANCEL_ADHOC_PROFILE,
+  SET_ADHOC_LEFT_PROFILE,
+  SET_ADHOC_RIGHT_PROFILE,
+  REQUEST_ADHOC_LEFT_PROFILE,
+  REQUEST_ADHOC_RIGHT_PROFILE,
+  RECEIVE_ADHOC_LEFT_PROFILE,
+  RECEIVE_ADHOC_RIGHT_PROFILE,
+  CANCEL_ADHOC_LEFT_PROFILE,
+  CANCEL_ADHOC_RIGHT_PROFILE,
+  REQUEST_ADHOC_PROFILE_DIFF,
+  RECEIVE_ADHOC_PROFILE_DIFF,
+  CANCEL_ADHOC_PROFILE_DIFF,
 } from './actionTypes';
 import { isAbortError } from '../util/abort';
+import { addNotification } from './reducers/notifications';
 
 export const setDateRange = (from, until) => ({
   type: SET_DATE_RANGE,
@@ -84,14 +106,17 @@ export const receiveTimeline = (data) => ({
   payload: data,
 });
 
-export const requestPyrescopeAppData = (url) => ({
-  type: REQUEST_PYRESCOPE_APP_DATA,
+export const requestPyroscopeAppData = (url) => ({
+  type: REQUEST_PYROSCOPE_APP_DATA,
   payload: { url },
 });
 
-export const receivePyrescopeAppData = (data) => ({
-  type: RECEIVE_PYRESCOPE_APP_DATA,
+export const receivePyroscopeAppData = (data) => ({
+  type: RECEIVE_PYROSCOPE_APP_DATA,
   payload: { data },
+});
+export const cancelPyroscopeAppData = () => ({
+  type: CANCEL_PYROSCOPE_APP_DATA,
 });
 
 export const requestComparisonAppData = (url, viewSide) => ({
@@ -103,6 +128,9 @@ export const receiveComparisonAppData = (data, viewSide) => ({
   type: RECEIVE_COMPARISON_APP_DATA,
   payload: { data, viewSide },
 });
+export const cancelComparisonappData = () => ({
+  type: CANCEL_COMPARISON_APP_DATA,
+});
 
 export const requestComparisonDiffAppData = (url) => ({
   type: REQUEST_COMPARISON_DIFF_APP_DATA,
@@ -112,6 +140,10 @@ export const requestComparisonDiffAppData = (url) => ({
 export const receiveComparisonDiffAppData = (data) => ({
   type: RECEIVE_COMPARISON_DIFF_APP_DATA,
   payload: { data },
+});
+
+export const cancelComparisonDiffAppData = () => ({
+  type: CANCEL_COMPARISON_DIFF_APP_DATA,
 });
 
 export const requestTags = () => ({ type: REQUEST_TAGS });
@@ -143,20 +175,139 @@ export const setQuery = (query) => ({
   payload: { query },
 });
 
-export const setFile = (file, flamebearer) => ({
-  type: SET_FILE,
+export const setAdhocFile = (file, flamebearer) => ({
+  type: SET_ADHOC_FILE,
   payload: { file, flamebearer },
 });
 
-export const setLeftFile = (file, flamebearer) => ({
-  type: SET_LEFT_FILE,
+export const setAdhocLeftFile = (file, flamebearer) => ({
+  type: SET_ADHOC_LEFT_FILE,
   payload: { file, flamebearer },
 });
 
-export const setRightFile = (file, flamebearer) => ({
-  type: SET_RIGHT_FILE,
+export const setAdhocRightFile = (file, flamebearer) => ({
+  type: SET_ADHOC_RIGHT_FILE,
   payload: { file, flamebearer },
 });
+
+export const requestAdhocProfiles = () => ({ type: REQUEST_ADHOC_PROFILES });
+
+export const receiveAdhocProfiles = (profiles) => ({
+  type: RECEIVE_ADHOC_PROFILES,
+  payload: { profiles },
+});
+
+export const cancelAdhocProfiles = () => ({ type: CANCEL_ADHOC_PROFILES });
+
+export const setAdhocProfile = (profile) => ({
+  type: SET_ADHOC_PROFILE,
+  payload: { profile },
+});
+
+export const requestAdhocProfile = () => ({ type: REQUEST_ADHOC_PROFILE });
+
+export const receiveAdhocProfile = (flamebearer) => ({
+  type: RECEIVE_ADHOC_PROFILE,
+  payload: { flamebearer },
+});
+
+export const cancelAdhocProfile = () => ({ type: CANCEL_ADHOC_PROFILE });
+
+export const setAdhocLeftProfile = (profile) => ({
+  type: SET_ADHOC_LEFT_PROFILE,
+  payload: { profile },
+});
+
+export const requestAdhocLeftProfile = () => ({
+  type: REQUEST_ADHOC_LEFT_PROFILE,
+});
+
+export const receiveAdhocLeftProfile = (flamebearer) => ({
+  type: RECEIVE_ADHOC_LEFT_PROFILE,
+  payload: { flamebearer },
+});
+
+export const cancelAdhocLeftProfile = () => ({
+  type: CANCEL_ADHOC_LEFT_PROFILE,
+});
+
+export const setAdhocRightProfile = (profile) => ({
+  type: SET_ADHOC_RIGHT_PROFILE,
+  payload: { profile },
+});
+
+export const requestAdhocRightProfile = () => ({
+  type: REQUEST_ADHOC_RIGHT_PROFILE,
+});
+
+export const receiveAdhocRightProfile = (flamebearer) => ({
+  type: RECEIVE_ADHOC_RIGHT_PROFILE,
+  payload: { flamebearer },
+});
+
+export const cancelAdhocRightProfile = () => ({
+  type: CANCEL_ADHOC_RIGHT_PROFILE,
+});
+
+export const requestAdhocProfileDiff = () => ({
+  type: REQUEST_ADHOC_PROFILE_DIFF,
+});
+
+export const receiveAdhocProfileDiff = (flamebearer) => ({
+  type: RECEIVE_ADHOC_PROFILE_DIFF,
+  payload: { flamebearer },
+});
+
+export const cancelAdhocProfileDiff = () => ({
+  type: CANCEL_ADHOC_PROFILE_DIFF,
+});
+
+// ResponseNotOkError refers to when request is not ok
+// ie when status code is not in the 2xx range
+class ResponseNotOkError extends Error {
+  constructor(response, text) {
+    super(`Bad Response with code ${response.status}: ${text}`);
+    this.name = 'ResponseNotOkError';
+    this.response = response;
+  }
+}
+
+// dispatchNotificationByError dispatches a notification
+// depending on the error passed
+function handleError(dispatch, e) {
+  if (e instanceof ResponseNotOkError) {
+    dispatch(
+      addNotification({
+        title: 'Request Failed',
+        message: e.message,
+        type: 'danger',
+      })
+    );
+  } else if (!isAbortError(e)) {
+    // AbortErrors are fine
+
+    // Generic case, so we use as message whatever error we got
+    // It's not the best UX, but our users should be experienced enough
+    // to be able to decipher what's going on based on the message
+    dispatch(
+      addNotification({
+        title: 'Error',
+        message: e.message,
+        type: 'danger',
+      })
+    );
+  }
+}
+
+// handleResponse retrieves the JSON data on success or raises an ResponseNotOKError otherwise
+function handleResponse(dispatch, response) {
+  if (response.ok) {
+    return response.json();
+  }
+  return response.text().then((text) => {
+    throw new ResponseNotOkError(response, text);
+  });
+}
 
 /**
  * ATTENTION! There may be race conditions:
@@ -182,16 +333,9 @@ export function fetchTimeline(url) {
     return fetch(`${url}&format=json`, {
       signal: currentTimelineController.signal,
     })
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receiveTimeline(data));
-      })
-      .catch((e) => {
-        // AbortErrors are fine
-        if (!isAbortError(e)) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveTimeline(data)))
+      .catch((e) => handleError(dispatch, e))
       .finally();
   };
 }
@@ -236,40 +380,28 @@ export function fetchComparisonAppData(url, viewSide) {
     return fetch(`${url}&format=json`, {
       signal: timelineController.signal,
     })
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receiveComparisonAppData(data, viewSide));
-      })
-      .catch((e) => {
-        // AbortError are fine
-        if (!isAbortError(e)) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveComparisonAppData(data, viewSide)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelComparisonappData()))
       .finally();
   };
 }
 
-export function fetchPyrescopeAppData(url) {
+export function fetchPyroscopeAppData(url) {
   return (dispatch) => {
     if (currentTimelineController) {
       currentTimelineController.abort();
     }
     currentTimelineController = new AbortController();
-    dispatch(requestPyrescopeAppData(url));
+    dispatch(requestPyroscopeAppData(url));
     return fetch(`${url}&format=json`, {
       signal: currentTimelineController.signal,
     })
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receivePyrescopeAppData(data));
-      })
-      .catch((e) => {
-        // AbortErrors are fine
-        if (!isAbortError(e)) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receivePyroscopeAppData(data)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelPyroscopeAppData()))
       .finally();
   };
 }
@@ -284,16 +416,11 @@ export function fetchComparisonDiffAppData(url) {
     return fetch(`${url}&format=json`, {
       signal: currentTimelineController.signal,
     })
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receiveComparisonDiffAppData(data));
-      })
-      .catch((e) => {
-        // AbortErrors are fine
-        if (!isAbortError(e)) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveComparisonDiffAppData(data)))
+      .catch((e) => handleError(dispatch, e))
+      .catch((e) => dispatchNotificationByError(dispatch, e))
+      .then(() => dispatch(cancelComparisonDiffAppData()))
       .finally();
   };
 }
@@ -307,16 +434,9 @@ export function fetchTags(query) {
 
     dispatch(requestTags());
     return fetch(`./labels?query=${encodeURIComponent(query)}`)
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receiveTags(data));
-      })
-      .catch((e) => {
-        // AbortErrors are fine
-        if (!isAbortError(e)) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveTags(data)))
+      .catch((e) => handleError(dispatch, e))
       .finally();
   };
 }
@@ -338,20 +458,13 @@ export function fetchTagValues(query, tag) {
 
     dispatch(requestTagValues(tag));
     return fetch(
-      `/label-values?label=${encodeURIComponent(
+      `./label-values?label=${encodeURIComponent(
         tag
       )}&query=${encodeURIComponent(query)}`
     )
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receiveTagValues(data, tag));
-      })
-      .catch((e) => {
-        // AbortErrors are fine
-        if (!fetchTagValuesController.signal.aborted) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveTagValues(data, tag)))
+      .catch((e) => handleError(dispatch, e))
       .finally();
   };
 }
@@ -372,26 +485,147 @@ export function fetchNames() {
     currentNamesController = new AbortController();
 
     dispatch(requestNames());
-    return fetch('/label-values?label=__name__', {
+    return fetch('./label-values?label=__name__', {
       signal: currentNamesController.signal,
     })
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(receiveNames(data));
-      })
-      .catch((e) => {
-        // AbortErrors are fine
-        if (!isAbortError(e)) {
-          throw e;
-        }
-      })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveNames(data)))
+      .catch((e) => handleError(dispatch, e))
       .finally();
   };
 }
-export function abortFetchNames() {
+
+let adhocProfilesController;
+export function fetchAdhocProfiles() {
+  return (dispatch) => {
+    if (adhocProfilesController) {
+      adhocProfilesController.abort();
+    }
+
+    adhocProfilesController = new AbortController();
+    dispatch(requestAdhocProfiles());
+    return fetch('./api/adhoc/v1/profiles', {
+      signal: adhocProfilesController.signal,
+    })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveAdhocProfiles(data)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelAdhocProfiles()))
+      .finally();
+  };
+}
+export function abortFetchAdhocProfiles() {
   return () => {
-    if (abortFetchNames) {
-      abortFetchNames.abort();
+    if (adhocProfilesController) {
+      adhocProfilesController.abort();
+    }
+  };
+}
+
+let adhocProfileController;
+export function fetchAdhocProfile(profile) {
+  return (dispatch) => {
+    if (adhocProfileController) {
+      adhocProfileController.abort();
+    }
+
+    adhocProfileController = new AbortController();
+    dispatch(requestAdhocProfile());
+    return fetch(`./api/adhoc/v1/profile/${profile}`, {
+      signal: adhocProfileController.signal,
+    })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveAdhocProfile(data)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelAdhocProfile()))
+      .finally();
+  };
+}
+export function abortFetchAdhocProfile() {
+  return () => {
+    if (adhocProfileController) {
+      adhocProfileController.abort();
+    }
+  };
+}
+
+let adhocLeftProfileController;
+export function fetchAdhocLeftProfile(profile) {
+  return (dispatch) => {
+    if (adhocLeftProfileController) {
+      adhocLeftProfileController.abort();
+    }
+
+    adhocLeftProfileController = new AbortController();
+    dispatch(requestAdhocLeftProfile());
+    return fetch(`./api/adhoc/v1/profile/${profile}`, {
+      signal: adhocLeftProfileController.signal,
+    })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveAdhocLeftProfile(data)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelAdhocLeftProfile()))
+      .finally();
+  };
+}
+export function abortFetchAdhocLeftProfile() {
+  return () => {
+    if (adhocLeftProfileController) {
+      adhocLeftProfileController.abort();
+    }
+  };
+}
+
+let adhocRightProfileController;
+export function fetchAdhocRightProfile(profile) {
+  return (dispatch) => {
+    if (adhocRightProfileController) {
+      adhocRightProfileController.abort();
+    }
+
+    adhocRightProfileController = new AbortController();
+    dispatch(requestAdhocRightProfile());
+    return fetch(`./api/adhoc/v1/profile/${profile}`, {
+      signal: adhocRightProfileController.signal,
+    })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveAdhocRightProfile(data)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelAdhocRightProfile()))
+      .finally();
+  };
+}
+export function abortFetchAdhocRightProfile() {
+  return () => {
+    if (adhocRightProfileController) {
+      adhocRightProfileController.abort();
+    }
+  };
+}
+
+let adhocProfileDiffController;
+export function fetchAdhocProfileDiff(left, right) {
+  return (dispatch) => {
+    if (adhocProfileDiffController) {
+      adhocProfileDiffController.abort();
+    }
+
+    adhocProfileDiffController = new AbortController();
+    dispatch(requestAdhocProfileDiff());
+    return fetch(`./api/adhoc/v1/diff/${left}/${right}`, {
+      signal: adhocProfileDiffController.signal,
+    })
+      .then((response) => handleResponse(dispatch, response))
+      .then((data) => dispatch(receiveAdhocProfileDiff(data)))
+      .catch((e) => handleError(dispatch, e))
+      .then(() => dispatch(cancelAdhocProfileDiff()))
+      .finally();
+  };
+}
+export function abortFetchAdhocProfileDiff() {
+  return () => {
+    if (adhocProfileDiffController) {
+      adhocProfileDiffController.abort();
     }
   };
 }